Change span_ends to prevent spilling
`span_ends` checks to see if the multiline span ends in this span. If I'm reading this correctly, it would be possible that the end of the span that ends at the previous line unintentionally spills into the next span, if that span starts at the beginning of the next line.

An example: if the offset is 10 and len is 2, we're ending at 12. I believe we don't want to consider the offset 12 to be part of this span, only a possible part of the next span.