fix(zero length): Improve rendering for zero-length error spans
A zero-length span at a line boundary can be associated either with the
previous or the next line. Current code prefer to use the next line,
however there isn't always a "next line" (end of source, or "no context"
configuration).
There is also the extra-special case of an empty document which has no
"next line" but also no "previous line".

This commit adds an empty newline at the end of a document if appropriate
so that there is a "next line", or use the "previous line"